requirements.txt
langchain
langchain-community
chromadb
pypdf
sentence-transformers